swdb: Don't create a new history DB for unprivileged user
When running dnf under an unprivileged user, do not check for the
presence of the history.sqlite file. The pathExists() check itself may
fail on isufficient permissions, but the function ignores this error and
behaves like the file doesn't exist. This breaks the logic that was in
place before this commit.

Instead, simply attempt to open the database file for reading and if
this fails, print the error and fall back to an in-memory database.

Creating a new database should not be possible for an unprivileged user
anyway, and all operations that modify the system (and therefore would
write to the history) are checking for the user being root too.

The new database will be empty and in that regard equivalent to an empty
in-memory database. Thew new empty database could potentially also
prevent conversion of formats of old databases if they are present.

This commit simply flips the pathExists() and getuid() conditions. For
the uid == 0 case the code is unchanged, for unprivileged user the
branch of the path not existing is dropped and the path existing branch
is again unchanged and used unconditionally.

Closes: #935
Approved by: <try>